update holo-cleent version 2.3.0

+++ holo-client/src/main/java/com/alibaba/hologres/client/impl/binlog/HoloBinlogDecoder.java
@@ -164,7 +164,7 @@ public class HoloBinlogDecoder {
 				}
 				break;
 			case Types.DATE:
-				currentRecord.setObject(index, new Date(currentRow.getLong(offsetIndex) * ONE_DAY_IN_MILLIES));
+				currentRecord.setObject(index, new Date(currentRow.getInt(offsetIndex) * ONE_DAY_IN_MILLIES));
 				break;
 			case Types.TIME:
 			case Types.TIME_WITH_TIMEZONE:
